requirements.txt
kafka-python
requests
grafana-client
apscheduler